requirements.txt
customtkinter>=5.2.0
pypdf>=3.0.0
pandas>=1.5.0
openpyxl>=3.1.0
openai>=1.0.0
Pillow>=9.0.0
requests>=2.28.0